services/requirements.txt
# ── Browser automation ────────────────────────────────────────────────────────
playwright>=1.44.0
playwright-stealth>=1.0.6

# ── HTML parsing ──────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=5.2.0                  # fast BS4 parser (optional but recommended)

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests>=2.32.0

# ── Search ────────────────────────────────────────────────────────────────────
typesense>=0.21.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0         # load .env for config
